feat(core): initial T1 dialogs
