+++ core/src/apps/common/confirm.py
@@ -1,6 +1,7 @@
 from trezor import wire
 from trezor.messages import ButtonRequestType
 from trezor.ui.confirm import CONFIRMED, INFO, Confirm, HoldToConfirm, InfoConfirm
+from trezor.ui.model import lookup_layout
 
 from . import button_request
 
@@ -9,7 +10,7 @@ if __debug__:
 
 
 if False:
-    from typing import Any, Callable, Optional
+    from typing import Any, Callable, Optional, Union, List, Dict
     from trezor import ui
     from trezor.ui.confirm import ButtonContent, ButtonStyleType
     from trezor.ui.loader import LoaderStyleType
@@ -112,5 +113,47 @@ async def require_confirm(*args: Any, **kwargs: Any) -> None:
 
 async def require_hold_to_confirm(*args: Any, **kwargs: Any) -> None:
     confirmed = await hold_to_confirm(*args, **kwargs)
+    if not confirmed:
+        raise wire.ActionCancelled
+
+
+# code using the functions above ^^^ should migrate to the interface below vvv
+
+# fmt: off
+_type_to_code = {
+    "confirm_backup1":                     ButtonRequestType.ResetDevice,
+    "confirm_backup2":                     ButtonRequestType.ResetDevice,
+    "confirm_change_count_over_threshold": ButtonRequestType.SignTx,
+    "confirm_feeoverthreshold":            ButtonRequestType.FeeOverThreshold,
+    "confirm_joint_total":                 ButtonRequestType.SignTx,
+    "confirm_nondefault_locktime":         ButtonRequestType.SignTx,
+    "confirm_output":                      ButtonRequestType.ConfirmOutput,
+    "confirm_path_warning":                ButtonRequestType.UnknownDerivationPath,
+    "confirm_ping":                        ButtonRequestType.ProtectCall,
+    "confirm_reset_device":                ButtonRequestType.ResetDevice,
+    "confirm_total":                       ButtonRequestType.SignTx,
+    "confirm_wipe":                        ButtonRequestType.WipeDevice,
+    "show_address":                        ButtonRequestType.Address,
+    "show_qr":                             ButtonRequestType.Address,
+    "warn_loading_seed":                   ButtonRequestType.Other,
+}  # type: Dict[str, EnumTypeButtonRequestType]
+# fmt: on
+
+
+async def interact(
+    ctx: wire.GenericContext, brtype: str, **content: Union[str, List[str]],
+) -> bool:
+    code = _type_to_code.get(
+        brtype, ButtonRequestType.Other
+    )  # FIXME: throw exception instead?
+    await button_request(ctx, code=code)
+
+    dialog = lookup_layout(brtype, content)
+
+    return await ctx.wait(dialog) is CONFIRMED
+
+
+async def require_interact(*args: Any, **kwargs: Any) -> None:
+    confirmed = await interact(*args, **kwargs)
     if not confirmed:
         raise wire.ActionCancelled
